# Extend the internal interface of the Blowfish module.