Hi,

I'm currently implementing bcrypt_pbkdf (as used by OpenSSH to encrypt private key files)
and found myself in need of certain low-level operations that are currently not exposed
by the Blowfish module.
After reading the code thoroughly I figured out how it has already been extended
in order to work for the regular BCrypt algorithm, but the eksBlowfish function is not general enough
and I wanted to avoid to just add another couple of extensions to the module instead
of trying to properly generalizing it and splitting separations of concern.
The new module interface is rather similar to that of other implementations, especially the Golang one.
No public interfaces or test cases had to be touched for this change.

I'm just seeking for a little input and feedback on this. I don't expect it to merged right away,
but just want to know whether such a change is generally welcome or rather too drastical.

P.S: This commit improves the performance of the BCrypt calculation by about 11% on my machine.

From the commit message:

In preparation of an implementation of the bcrypt_pbkdf (a
variant of PBKDF2 used by OpenSSH) algorithm,
certain low-level operations of the Blowfish algorithm need to
be generalized and exposed.

The Blowfish.Primitive module has already been extended to
account for the requirements imposed by the BCrypt algorithm,
but the salt length was limited to 16 bytes and the BCrypt
specific key schedule setup has been hard-coded into the Blowfish
module.

This commit makes a clear distintion between the expandKey and
expandKeyWithSalt operation. Both take arbitrary sized salts
and keys now. The specialized operation for 16 byte salts as used
by BCrypt has been preserved and is selected automatically.
Also, the BCrypt specific parts have been move to the BCrypt
module with regard to separation of concern.

A benchmark for generating BCrypt hashes with cost 10 shows a
performance improvement from 158 to 141ms on average (Intel i5-6500)
after this refactoring.
Further experiments suggest that the specialized expandKeyWithSalt128
does not have any advantage over the generalized version
and might be removed in favour of less branches and exceptional
behaviour.