@ocheron 2019-03-03
The transformed implementation looks good.

For BCrypt with GHC 8.6.3, performance is further improved when adding a bang pattern to the first argument `i` of `doRound` inner functions. By something like 30%, not only 11%. And for Blowfish128 the PR gives a good performance improvement too.

Before:
```
block-cipher/ECB/Blowfish128-input=1024  mean 51.97 μs  ( +- 764.5 ns  )
bcrypt/cryptonite-BCrypt-4               mean 2.566 ms  ( +- 21.72 μs  )
bcrypt/cryptonite-BCrypt-5               mean 4.949 ms  ( +- 27.13 μs  )
bcrypt/cryptonite-BCrypt-7               mean 19.21 ms  ( +- 76.61 μs  )
bcrypt/cryptonite-BCrypt-11              mean 304.3 ms  ( +- 595.1 μs  )
```

After (PR):
```
block-cipher/ECB/Blowfish128-input=1024  mean 16.48 μs  ( +- 161.8 ns  )
bcrypt/cryptonite-BCrypt-4               mean 2.265 ms  ( +- 14.05 μs  )
bcrypt/cryptonite-BCrypt-5               mean 4.430 ms  ( +- 50.78 μs  )
bcrypt/cryptonite-BCrypt-7               mean 17.35 ms  ( +- 103.6 μs  )
bcrypt/cryptonite-BCrypt-11              mean 275.3 ms  ( +- 338.0 μs  )
```

After (PR modified with bang pattern):
```
block-cipher/ECB/Blowfish128-input=1024  mean 16.46 μs  ( +- 102.1 ns  )
bcrypt/cryptonite-BCrypt-4               mean 1.711 ms  ( +- 9.857 μs  )
bcrypt/cryptonite-BCrypt-5               mean 3.324 ms  ( +- 19.37 μs  )
bcrypt/cryptonite-BCrypt-7               mean 13.01 ms  ( +- 75.34 μs  )
bcrypt/cryptonite-BCrypt-11              mean 206.2 ms  ( +- 308.7 μs  )
```